fix: absolute paths in build info
